Refactor URL_BASE constant in Utils.java, add Usuario role in Roles.java, add modelmapper dependency in pom.xml, add constructors in ZonaDto.java, rename creatRele to createRole in IRoleService.java, add existsByFileName method in ISupportDocumentsRepository.java, delete GanderoDto.java, change JpaRepository generic type to Long in IRoleRepository.java, add HttpStatus parameter in CustomerNotExistException.java, change Date type to String and add accion field in HistoryAuditordDto.java, add updateHistorial method in IHistorialAuditoriaService.java

+++ src/main/java/com/alcadia/bovid/Repository/Dao/IRoleRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.alcadia.bovid.Models.Entity.Role;
 
 @Repository
-public interface IRoleRepository  extends JpaRepository<Role, Integer>{
+public interface IRoleRepository  extends JpaRepository<Role,Long>{
     Optional<Role> findByAuthority(String authority);
 
     Void deleteByAuthority(String authority);
